backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.11
cryptography==41.0.7
alembic==1.13.0

# ML and Data Processing
torch==2.9.1
scikit-learn==1.5.2
joblib==1.4.2
numpy==1.26.4
pandas==2.2.3

# Spatial/GIS
geoalchemy2==0.14.3
shapely==2.0.2

# OpenAI Integration
openai==1.3.7

# Replicate API for AI models
replicate>=0.25.0

# Computer Vision
opencv-python==4.10.0.84

# Redis for job queue
redis==5.0.1
celery==5.3.4

# File handling
openpyxl==3.1.2
Pillow==11.0.0

# Utilities
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development
black==23.12.0
flake8==6.1.0
mypy==1.7.1
